# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "My Barber API",
    "version": "1.0.0",
    "description": "Barber booking API built with Node.js and TypeScript.\n\n**Schemas:** `components.schemas` mirrors TypeScript types in `models/` (e.g. `BookingStatus`, `BookingContract` ↔ `models/booking.ts`; barber approval ↔ `models/barber.ts`). Enumerations are OpenAPI `enum`s; request/response bodies use `$ref` where noted.",
    "contact": {
      "name": "API Support",
      "email": "support@my-barber.uz"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.my-barber.uz",
      "description": "Production server"
    },
    {
      "url": "https://staging-api.my-barber.uz",
      "description": "Staging server"
    },
    {
      "url": "http://localhost:8080",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Barber/client JWT from login endpoints, or a Firebase Auth ID token with custom claim admin:true / role:admin for Admin routes. Dev/bootstrap: JWT whose user `id` is in ADMIN_UID_ALLOWLIST."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Error message"
          },
          "details": {
            "type": "array",
            "description": "Optional express-validator errors (ValidationError[] style).",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      },
...
      "ReportTargetType": {
        "type": "string",
        "enum": [
          "barber",
          "client",
          "booking",
          "review"
        ]
      },
...
      "ReportCreateRequest": {
        "type": "object",
        "required": [
          "target_type",
          "target_id",
          "reason"
        ],
        "properties": {
          "target_type": {
            "$ref": "#/components/schemas/ReportTargetType"
          },
          "target_id": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "attachments": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "BlockCreateRequest": {
        "type": "object",
        "required": [
          "blocked_user_id"
        ],
        "properties": {
          "blocked_user_id": {
            "type": "string"
          }
        }
      },
...
      "ReportCreateResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "properties": {
              "report": {
                "type": "object",
                "additionalProperties": true,
                "description": "Persisted report document"
              }
            }
          }
        }
      },
      "EmptyDataEnvelope": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "data": {
            "type": "object"
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/reports": {
      "post": {
        "tags": [
          "Trust"
        ],
        "summary": "Submit abuse/safety report",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Report created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportCreateResponse"
                }
              }
            }
          }
        }
      }
    },
    "/blocks": {
      "post": {
        "tags": [
          "Trust"
        ],
        "summary": "Block another user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyDataEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Invalid block",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/blocks/{userId}": {
      "delete": {
        "tags": [
          "Trust"
        ],
        "summary": "Unblock user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyDataEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Block not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}